requirements.txt
datasets
arxiv
langgraph 
langchain_community 
langchain-google-community
langchain_anthropic
langchain_mongodb
langchain_openai
pymongo[srv]
python-dotenv
pydantic-core
pydantic
pymupdf
langsmith 
openai
unstructured
networkx
pandas
streamlit
streamlit-file-browser

